fix(frontend): use existing invite onboarding flow
