+++ src/modules/auth.ts
@@ -108,25 +108,6 @@ async function updateUser(
   }
 }
 
-async function hasPendingInvitations(supabase: SupabaseClient) {
-  try {
-    const { data, error } = await supabase.functions.invoke('private/pending_invitations', {
-      method: 'GET',
-    })
-
-    if (error) {
-      console.error('Failed to load pending organization invitations', error)
-      return false
-    }
-
-    return (data?.invitations?.length ?? 0) > 0
-  }
-  catch (error) {
-    console.error('Failed to load pending organization invitations', error)
-    return false
-  }
-}
-
 async function maybeProvisionSsoMembership(
   supabase: SupabaseClient,
   session: Awaited<ReturnType<SupabaseClient['auth']['getSession']>>['data']['session'] | null,
@@ -230,12 +211,12 @@ async function guard(
     return !organizationStore.organizations.some(org => org.gid === inviteOrgId && org.role.startsWith('invite'))
   }
 
-  async function shouldRedirectToPendingInviteOnboarding(organizationsLoaded: boolean) {
+  function shouldRedirectToPendingInviteOnboarding(organizationsLoaded: boolean) {
     if (!organizationsLoaded)
       return false
     if (to.path.startsWith('/onboarding/invitation'))
       return false
-    return await hasPendingInvitations(supabase)
+    return organizationStore.organizations.some(org => org.role.startsWith('invite'))
   }
 
   if (hasAuth && sessionUser) {
@@ -295,7 +276,7 @@ async function guard(
     }
 
     const organizationsLoaded = await tryLoadOrganizations(() => organizationStore.fetchOrganizations())
-    if (await shouldRedirectToPendingInviteOnboarding(organizationsLoaded)) {
+    if (shouldRedirectToPendingInviteOnboarding(organizationsLoaded)) {
       return next({
         path: '/onboarding/invitation',
         query: {
@@ -370,7 +351,7 @@ async function guard(
     }
 
     let organizationsLoaded = await tryLoadOrganizations(() => organizationStore.dedupFetchOrganizations())
-    if (await shouldRedirectToPendingInviteOnboarding(organizationsLoaded)) {
+    if (shouldRedirectToPendingInviteOnboarding(organizationsLoaded)) {
       return next({
         path: '/onboarding/invitation',
         query: {
